add RandomRecipesTable component to divide code and make components shorter extract weekdays into separate file
