+++ src/modules/weekdays.js
@@ -1,0 +1,11 @@
+const weekdays = [
+  "Monday",
+  "Tuesday",
+  "Wednesday",
+  "Thursday",
+  "Friday",
+  "Saturday",
+  "Sunday"
+];
+
+export default weekdays;
